chore(engine): remove unnecessary unboxing

+++ engine/src/main/java/org/camunda/bpm/engine/impl/history/producer/DefaultDmnHistoryEventProducer.java
@@ -113,7 +113,7 @@ public class DefaultDmnHistoryEventProducer implements DmnHistoryEventProducer {
       return ((LongValue) collectResultValue).getValue().doubleValue();
 
     } else if(collectResultValue instanceof DoubleValue) {
-      return ((DoubleValue) collectResultValue).getValue().doubleValue();
+      return ((DoubleValue) collectResultValue).getValue();
 
     } else {
       throw LOG.collectResultValueOfUnsupportedTypeException(collectResultValue);
